requirements.txt
dash==2.8.0
dash_ag_grid==31.2.0
dash_bootstrap_components==1.6.0
dash_iconify==0.1.2
dash_mantine_components==0.12.0
pandas==2.0.2
parsel==1.8.1
price_parser==0.3.4
python-dotenv==1.0.1
requests==2.31.0
gunicorn
dash-tools
